requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
